packages/ai-agent/requirements.txt
google-adk
requests
requests_oauthlib
fastapi
uvicorn
httpx
python-dotenv
pydantic